+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Changed
-- Client version updated on [5.0.16](https://github.com/reportportal/client-java/releases/tag/5.0.16)
+- Client version updated on [5.0.17](https://github.com/reportportal/client-java/releases/tag/5.0.17)
 ### Fixed
 - Empty interrupted suite in case of duplicate step