+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Changed
-- Client version updated on [5.0.18](https://github.com/reportportal/client-java/releases/tag/5.0.18)
+- Client version updated on [5.0.21](https://github.com/reportportal/client-java/releases/tag/5.0.21)
 ### Fixed
 - Empty interrupted suite in case of duplicate step