+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [5.1.0-RC-1]
+
 ## [5.0.2]
 ### Changed
 - Client version updated on [5.0.21](https://github.com/reportportal/client-java/releases/tag/5.0.21)